pyproject.toml
[project]
name = "ai-laywer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.34",
    "langchain-groq>=0.2.4",
    "langchain-ollama>=0.2.3",
    "langchain-text-splitters>=0.3.6",
    "pdfplumber>=0.11.5",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
    "watchdog>=6.0.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]